Changes/cleanups/tweaks required to build samba.
1) Move the ENDIAN macros to inttypes.h so they get picked since nobody
   includes endian.h directly.
2) Fix several prototypes
3) Fix some member names in siginfo_t
4) Add IP_TOS
5) Take out the *Stat() macros. This was colliding with a variable name in
   the samba code. Instead, let libc_nonshared.a get used now that it's use
   has been straightened out in lsbcc.

+++ headers/sys/ioctl.h
@@ -15,5 +15,5 @@
 
 
 
-extern int ioctl (int, int, ...);
+extern int ioctl (int, unsigned long, ...);
 #endif
